Create WithdrawableTreeOperations for common ops
The new WithdrawableTreeOperations contains functions that apply regardless of the entity type being withdrawn/cascaded. This is in preperation for adding more type-specific operations to Withdrawal Service to trigger the top level withdrawal processes.

Moving the generic withdrawDescendantsOfRootNode function into a standalone class will make the unit tests and classes easier to manage. This provided a good opportunituy to simplify the unit tests previously in withdrawal service that were essentially testing the same withdrawal functionality with different depths of trees.
